pyproject.toml
[project]
name = "data_guide"
version = "0.0.1"
description = "Accelerated and Augmented EDA"
authors = [{ name = "Zach" }]
requires-python = ">=3.11"
dependencies = [
  "pandas",
  "numpy",
  "scipy",
  "matplotlib",
  "seaborn",
  "missingno",
  "phonenumbers",
  "statsmodels",
  "scikit-learn",
  "geopandas",
  "esda",
  "requests",
  "beautifulsoup4",
  "python-magic",
  "markdown",
  "pdfplumber",
  "typer",
  "pyyaml",
  "tabulate>=0.9.0",
]

[project.scripts]
guide = "typer_cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pyyaml>=6.0.2",
    "ruff>=0.13.1",
]
